pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apibackuper"
version = "1.0.11"
description = "a command-line tool and python library for API backuping"
readme = "README.md"
requires-python = ">=3.6"
license = {text = "MIT"}
authors = [
    {name = "Ivan Begtin", email = "ivan@begtin.tech"}
]
keywords = ["api", "json", "jsonl", "csv", "bson", "cli", "dataset", "backup", "archive"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Topic :: Software Development",
    "Topic :: System :: Networking",
    "Topic :: Terminals",
    "Topic :: Text Processing",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.9.0",
    "lxml>=4.9.3",
    "urllib3>=2.0.6",
    "requests>=2.31.0",
    "xmltodict>=0.13.0",
    "PyYAML>=6.0",
    "jsonschema>=4.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
# Optional but recommended dependencies
recommended = [
    "aria2p>=0.11.3",
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
]
dev = [
    "pre-commit>=3.6.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "mock>=5.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "flake8-docstrings>=1.7.0",
    "flake8-bugbear>=23.12.0",
    "flake8-comprehensions>=3.14.0",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
    "bandit>=1.7.6",
    "pylint>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/datacoon/apibackuper/"
Documentation = "https://github.com/datacoon/apibackuper/"
Repository = "https://github.com/datacoon/apibackuper/"
Issues = "https://github.com/datacoon/apibackuper/issues"
Download = "https://github.com/datacoon/apibackuper/"

[project.scripts]
apibackuper = "apibackuper.__main__:main"

[tool.setuptools]
packages = ["apibackuper"]
include-package-data = true
zip-safe = false

[tool.setuptools.package-data]
apibackuper = ["schemas/*.json"]

# Black code formatter configuration
[tool.black]
line-length = 100
target-version = ["py36", "py37", "py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | apibackuper\.egg-info
)/
'''

# isort import sorter configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = [
    "build/*",
    "dist/*",
    ".eggs/*",
    "*.egg-info/*",
    ".tox/*",
    ".venv/*",
]

# mypy type checker configuration
[tool.mypy]
python_version = "3.6"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = true
strict_optional = false
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "lxml.*",
    "xmltodict.*",
    "aria2p.*",
]
ignore_missing_imports = true

# pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--doctest-modules",
    "--cov=apibackuper",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=0",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["apibackuper"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/setup.py",
    "*/build/*",
    "*/.tox/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Bandit security linter configuration
[tool.bandit]
exclude_dirs = ["tests", "test_*.py"]
skips = ["B101", "B601"]  # Skip assert_used and shell_injection_subprocess

# Flake8 configuration (also in setup.cfg, but keeping here for completeness)
[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "E501"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg",
    ".tox",
    ".venv",
    ".mypy_cache",
]

